Copy parametrize

+++ src/DB/Transfer.php
@@ -51,13 +51,178 @@ class Transfer {
         }
     }
 
-    private function processFields(&$row) {
+    /**
+     * Processes each field if necessary by handler at params array and PDO quote
+     * @param array $row
+     * @param array $columnsMap [oldFieldName => newFieldName]
+     * @param array $params [handler => callable, sourceColumns => array]
+     * @return array
+     */
+    private function processFields($row, $columnsMap, $params) {
+        $newRow = array();
         foreach($row as $name => &$field) {
-            $field = "'{$field}'";
+            if (isset($columnsMap[$name])) {
+                $newName = $columnsMap[$name];
+                if (isset($params['handler'])) {
+                    $type = isset($params['sourceColumns'][$name]['type']) ? $params['sourceColumns'][$name]['type'] : false;
+                    $field = call_user_func($params['handler'], $name, $type, $field);
+                }
+                $newRow[$newName] = $this->toDB->quote($field);
+            }
         }
-        return $row;
+        return $newRow;
     }
 
+    /**
+     * Prepare columns for create new table
+     * @param array $columns
+     * @param array $params
+     * @return mixed
+     */
+    public function processColumns($columns, $params) {
+        $rules = array('default' => 'default', 'whitelist' => 'whitelist', 'blacklist' => 'blacklist');
+        if (isset($params['fields'])) {
+            $fieldsRule = isset($params['fields'][0]) && isset($rules[strtolower($params['fields'][0])]) ?
+                $rules[strtolower($params['fields'][0])] : reset($rules);
+            unset($params['fields'][0]);
+        }
+
+        if (isset($params['types'])) {
+            $typesRule = isset($params['types'][0]) && isset($rules[strtolower($params['types'][0])]) ?
+                $rules[strtolower($params['types'][0])] : reset($rules);
+            unset($params['types'][0]);
+        }
+
+        foreach ($columns as $name => $column) {
+            if (isset($fieldsRule)) {
+                if ('whitelist' === $fieldsRule) {
+                    if (!isset($params['fields'][$name])) {
+                        unset($columns[$name]);
+                    } elseif ($params['fields'][$name] != $name) {
+                        $columns[$params['fields'][$name]] = $columns[$name];
+                        unset($columns[$name]);
+                    }
+                } elseif('blacklist' === $fieldsRule) {
+                    if (in_array($name, $params['fields'])) {
+                        unset($columns[$name]);
+                    }
+                } elseif ($params['fields'][$name] != $name) {
+                    $columns[$params['fields'][$name]] = $columns[$name];
+                    unset($columns[$name]);
+                }
+            }
+
+            if (isset($typesRule)) {
+                if ('whitelist' === $typesRule) {
+                    $found = false;
+                    foreach ($params['types'] as $oldType => $newType) {
+                        if (strstr($column['type'], $oldType)) {
+                            $found = true;
+                            $columns[$name]['type'] = $newType;
+                            break;
+                        }
+                    }
+                    if (!$found) {
+                        unset($columns[$name]);
+                    }
+                } elseif('blacklist' === $typesRule) {
+                    $found = false;
+                    foreach ($params['types'] as $oldType => $newType) {
+                        if (strstr($column['type'], $oldType)) {
+                            $found = true;
+                            break;
+                        }
+                    }
+                    if ($found) {
+                        unset($columns[$name]);
+                    }
+                } else {
+                    foreach ($params['types'] as $oldType => $newType) {
+                        if (strstr($column['type'], $oldType)) {
+                            $columns[$name]['type'] = $newType;
+                            break;
+                        }
+                    }
+                }
+            }
+        }
+
+        return $columns;
+    }
+
+    /**
+     * Creates array of columns map, that consist of pairs oldFieldName => newFieldName
+     * @param array $columns
+     * @param array $params
+     * @return array
+     */
+    public function createColumnsMap($columns, $params) {
+        $columnsMap = array();
+
+        $rules = array('default' => 'default', 'whitelist' => 'whitelist', 'blacklist' => 'blacklist');
+        if (isset($params['fields'])) {
+            $fieldsRule = isset($params['fields'][0]) && isset($rules[strtolower($params['fields'][0])]) ?
+                $rules[strtolower($params['fields'][0])] : reset($rules);
+            unset($params['fields'][0]);
+        }
+
+        if (isset($params['types'])) {
+            $typesRule = isset($params['types'][0]) && isset($rules[strtolower($params['types'][0])]) ?
+                $rules[strtolower($params['types'][0])] : reset($rules);
+            unset($params['types'][0]);
+        }
+
+        foreach ($columns as $name => $column) {
+            if (isset($fieldsRule)) {
+                if ('whitelist' === $fieldsRule) {
+                    if (isset($params['fields'][$name])) {
+                        $columnsMap[$name] = $params['fields'][$name];
+                    }
+                } elseif('blacklist' === $fieldsRule) {
+                    if (!in_array($name, $params['fields'])) {
+                        $columnsMap[$name] = $name;
+                    }
+                } else {
+                    $columnsMap[$name] = $params['fields'][$name];
+                }
+            }
+
+            if (isset($typesRule)) {
+                if ('whitelist' === $typesRule) {
+                    $found = false;
+                    foreach ($params['types'] as $oldType => $newType) {
+                        if (strstr($column['type'], $oldType)) {
+                            $found = true;
+                            break;
+                        }
+                    }
+                    if ($found) {
+                        $columnsMap[$name] = $name;
+                    }
+                } elseif('blacklist' === $typesRule) {
+                    $found = false;
+                    foreach ($params['types'] as $oldType => $newType) {
+                        if (strstr($column['type'], $oldType)) {
+                            $found = true;
+                            break;
+                        }
+                    }
+                    if (!$found) {
+                        $columnsMap[$name] = $name;
+                    }
+                }
+            }
+        }
+
+        return $columnsMap;
+    }
+
+    /**
+     * Copy tables from fromDB to toDB
+     * @param string|array $tables
+     * @param array $params
+     * @throws Exception
+     */
     public function copy($tables, $params = []) {
         $tables = !is_array($tables) ? [$tables] : $tables;
         foreach ($tables as $tableName => $tableParams) {
@@ -78,19 +243,26 @@ class Transfer {
                 throw new Exception('Incorrect tables');
             }
 
+            if (is_array($tableParams)) {
+                $params = array_merge($tableParams, $params);
+            }
+
             if ((false !== $fromTable) && !$this->fromDB->isTableExists($fromTable)) {
                 throw new Exception('Table doesn\'t exists at first table');
             }
 
             $select = isset($select) ? $select : $this->fromDB->getSQLBuilder()->from($fromTable);
 
+            $columns = $this->processColumns($this->fromDB->getColumns($fromTable), $params);
+            $columnsMap = $this->createColumnsMap($this->fromDB->getColumns($fromTable), $params);
+
             if ($this->toDB->isTableExists($toTable)) {
-                ///*if (!$this->toDB->isTableEqual($tableName, $this->fromDB->getColumns($tableName))) {
+                if (!$this->toDB->isTableEqual($toTable, $columns)) {
                     $this->toDB->dropTable($toTable);
-                    $this->toDB->createTable($toTable, $this->fromDB->getColumns($fromTable));
-                //}*/
+                    $this->toDB->createTable($toTable, $columns);
+                }
             } else {
-                $this->toDB->createTable($toTable, $this->fromDB->getColumns($fromTable));
+                $this->toDB->createTable($toTable, $columns);
             }
 
             $continue = true;
@@ -99,7 +271,7 @@ class Transfer {
                 $continue = false;
                 if ($rows = $this->fromDB->execute($select->limit($this->portion)->offset($offset)->getSQL())->fetchAll(BasePDO::FETCH_ASSOC)) {
                     foreach ($rows as $row) {
-                        $row = $this->processFields($row);
+                        $row = $this->processFields($row, $columnsMap, array_merge($params, array('sourceColumns' => $this->fromDB->getColumns($fromTable))));
                         $insertUpdate = $this->toDB->getSQLBuilder()->insertOnDuplicateUpdate($toTable, $row);
 
                         if (false === $this->toDB->execute($insertUpdate)) {
